Complete sign up form page
the signup form page is style to replicate the
blueprint provided in png. I'm pretty happy with it
all I would do to get a picture perfect static page now
would be to change the font-family, font scaling,
dummy-text, and border color on password invalid entries.
Otherwise, I'm pretty happy with it.